+++ src/game/scenes/pausescene.cc
@@ -10,17 +10,17 @@ namespace scene {
 
 PauseScene::PauseScene () {
     ugdk::graphic::SolidRectangle* rect = new ugdk::graphic::SolidRectangle(VIDEO_MANAGER()->video_size());
-    rect->set_color(ugdk::Color(0.5f, 0.5f, 0.5f));
+    rect->set_color(ugdk::Color(0.5, 0.5, 0.5));
 
     ugdk::graphic::Node* node = new ugdk::graphic::Node(rect);
-    node->modifier()->set_alpha(0.5f);
+    node->modifier()->set_alpha(0.5);
 
     this->root_node()->AddChild(node);
 }
 
 PauseScene::~PauseScene () {}
 
-void PauseScene::Update (float delta_t) {
+void PauseScene::Update (double delta_t) {
     if (INPUT_MANAGER()->KeyPressed(ugdk::input::K_RETURN))
         Finish();
 }
