+++ src/game/scenes/world.h
@@ -36,7 +36,7 @@ class World : public ugdk::Scene {
     World(sprite::Hero *hero, utils::ImageFactory *factory);
     virtual ~World();
 
-	void Update(float delta_t);
+	void Update(double delta_t);
 
     void AddWorldObject(sprite::WorldObject*, ugdk::Vector2D pos);
     void AddHero(ugdk::Vector2D pos);
@@ -64,7 +64,7 @@ class World : public ugdk::Scene {
     // Funcao que transforma PONTOS de coordenadas de tela para de mundo
     static Vector2D FromScreenCoordinates(Vector2D screen_coords);
 
-	static const Vector2D ConvertLightRadius(float radius);
+	static const Vector2D ConvertLightRadius(double radius);
 
     //getters
     sprite::Hero * hero() const { return hero_; }
@@ -96,7 +96,7 @@ class World : public ugdk::Scene {
 
     Vector2D ActualOffset();
     bool verifyCollision(sprite::WorldObject *obj1, sprite::WorldObject *obj2);
-    void VerifyCheats(float delta_t);
+    void VerifyCheats(double delta_t);
     bool VerifyPause();
     void HandleCollisions();
     void RemoveInactiveObjects();
