+++ src/game/sprites/creatures/pharaoh.h
@@ -30,8 +30,8 @@ class Pharaoh : public Mummy {
   protected:
 	skills::Skill *ranged_weapon_, *summon_weapon_;
 
-	void Update(float delta_t);
-    void Think(float dt);
+	void Update(double delta_t);
+    void Think(double dt);
 
 };
